+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   native C/C++ crash reporting if the current `ReleaseStage` is not in the array
 * Fix erroneous automatic session delivery when `NotifyReleaseStages`
   configuration does not include the current `ReleaseStage`
+* Fix the behavior of `Configuration.AutoNotify` to disable native crash
+  reporting if false
 * Update bugsnag-cocoa to v5.22.6:
   * Ensure UIKit APIs are not called from background threads if
     `Bugsnag.start()` is called in the background
@@ -23,7 +25,6 @@
     will be detected and sent. This interface can be used for crash reporting
     opt-out flows.
     [bugsnag-cocoa#410](https://github.com/bugsnag/bugsnag-cocoa/issues/410)
-  
 
 ## 4.6.4 (2019-09-06)